+++ src/main/java/joshie/harvest/core/helpers/MCClientHelper.java
@@ -1,8 +1,10 @@
 package joshie.harvest.core.helpers;
 
 import joshie.harvest.core.network.PenguinPacketLocation;
+import joshie.harvest.shops.gui.GuiNPCShop;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.entity.EntityPlayerSP;
+import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;