+++ src/main/java/joshie/harvest/quests/player/friendship/QuestTiberius.java
@@ -1,0 +1,21 @@
+package joshie.harvest.quests.player.friendship;
+
+import joshie.harvest.api.quests.HFQuest;
+import joshie.harvest.api.quests.Quest;
+import joshie.harvest.npc.HFNPCs;
+import joshie.harvest.quests.Quests;
+import joshie.harvest.quests.base.QuestFriendship;
+
+import java.util.Set;
+
+@HFQuest("friendship.tiberius.dark")
+public class QuestTiberius extends QuestFriendship {
+    public QuestTiberius() {
+        super(HFNPCs.CLOCKMAKER, 15000);
+    }
+
+    @Override
+    public boolean canStartQuest(Set<Quest> active, Set<Quest> finished) {
+        return finished.contains(Quests.RECIPE_DOUGHNUT);
+    }
+}