+++ src/main/java/joshie/harvest/quests/player/recipes/QuestPorridge.java
@@ -1,12 +1,21 @@
 package joshie.harvest.quests.player.recipes;
 
 import joshie.harvest.api.quests.HFQuest;
+import joshie.harvest.api.quests.Quest;
 import joshie.harvest.npc.HFNPCs;
+import joshie.harvest.quests.Quests;
 import joshie.harvest.quests.base.QuestRecipe;
+
+import java.util.Set;
 
 @HFQuest("recipe.porridge")
 public class QuestPorridge extends QuestRecipe {
     public QuestPorridge() {
         super("porridge", HFNPCs.CAFE_GRANNY, 5000);
+    }
+
+    @Override
+    public boolean canStartQuest(Set<Quest> active, Set<Quest> finished) {
+        return finished.contains(Quests.MEET_KATLIN);
     }
 }